pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.75",
    "langchain-google-genai>=2.1.10",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "tavily-python>=0.7.11",
    "uvicorn>=0.35.0",
]
